+++ src/count_min_sketch.rs
@@ -1,4 +1,4 @@
-use crate::utils::*;
+use crate::utils::helpers::*;
 use anyhow::{Context, Result};
 use murmur3::murmur3_x64_128;
 use rand::Rng;
